+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 
 - cli output `dateCreated` `addedDate` localize, use system TZ or given by ENV
 - login feature [#1](/../../issues/1)
+- last time, 2.3days , 3.2hours , 11minutes, etc
 
 ---------